+++ stories/02-Design System/06-Feedback/Modal/data.ts
@@ -12,6 +12,19 @@ export const argTypes = {
             category: 'Modal'
         }
     },
+    fullscreen: {
+        name: 'fullscreen',
+        description: 'Should the modal open fullscreen',
+        table: {
+            type: {
+                summary: 'boolean'
+            },
+            defaultValue: {
+                summary: false
+            },
+            category: 'Modal'
+        }
+    },
     style: {
         name: 'style',
         description: