+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## [x.x.x](https://www.npmjs.com/package/aristid-ds/v/x.x.x) (2024-xx-xx)
+
+### ✨ Features
+* **Modal:**
+    * Add `fullscreen`option (@evoiron)
+
 ### 📚 Documentation
 * New documentation page to list available design tokens (@evoiron)
 * Display each design token corresponding css variable (@evoiron)