+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### 📚 Documentation
 * New documentation page to list available design tokens (@evoiron)
 * Display each design token corresponding css variable (@evoiron)
+* Display package version number on the top right corner of Storybook (@philippechevieux)
 
 ## [0.11.0](https://www.npmjs.com/package/aristid-ds/v/0.11.0) (2024-01-24)
 ### ✨ Features